# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for AllianceCup Application",
        "title": "AllianceCup API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/client/add-to-cart": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "adds a product to a cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/client"
                ],
                "summary": "AddToCart",
                "operationId": "adds a product to a cart",
                "parameters": [
                    {
                        "description": "product info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/shopping.CartProduct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/api/client/add-to-favourites": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "adds a product to favourites",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/client"
                ],
                "summary": "AddToFavourites",
                "operationId": "adds to favourites",
                "parameters": [
                    {
                        "description": "product id",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ProductIdInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ItemProcessedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
...
        "/api/client/delete-from-cart": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "deletes a product from users cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/client"
                ],
                "summary": "DeleteFromCart",
                "operationId": "deletes from cart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ItemProcessedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/api/client/delete-from-favourites": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "deletes product from favourites",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/client"
                ],
                "summary": "DeleteFromFavourites",
                "operationId": "deletes from favourites",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.ItemProcessedResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/api/client/get-favourites": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "gets user favourite products",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/client"
                ],
                "summary": "GetFavourites",
                "operationId": "get favourites",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/product.Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
...
        "/api/client/user-cart": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "gets products from a cart",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/client"
                ],
                "summary": "GetProductsInCart",
                "operationId": "gets products from a cart",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.CartProductsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
...
        "/api/get-products": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "GetProducts",
                "operationId": "gets products",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category",
                        "name": "category",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Price",
                        "name": "price",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "characteristic",
                        "name": "characteristic",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Created At",
                        "name": "created_at",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/product.Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.CartProductsResponse": {
            "type": "object",
            "properties": {
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/shopping.CartProductFullInfo"
                    }
                },
                "sum": {
                    "type": "number"
                }
            }
        },
...
        "handler.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.ItemProcessedResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.ProductIdInput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "product.Product": {
            "type": "object",
            "required": [
                "article",
                "category_title",
                "packages_in_box",
                "price",
                "product_title",
                "type_title",
                "units_in_package"
            ],
            "properties": {
                "amount_in_stock": {
                    "type": "number",
                    "example": 120
                },
                "article": {
                    "type": "string",
                    "example": "000123"
                },
                "category_title": {
                    "type": "string",
                    "example": "Одноразові стакани"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 5
                },
                "img_url": {
                    "type": "string",
                    "example": "https://google-images.com/some-img123"
                },
                "packages_in_box": {
                    "type": "integer",
                    "example": 50
                },
                "price": {
                    "type": "number",
                    "example": 3.75
                },
                "product_title": {
                    "type": "string",
                    "example": "Стакан одноразовий Крафт 110мл"
                },
                "type_title": {
                    "type": "string",
                    "example": "Стакан"
                },
                "units_in_package": {
                    "type": "integer",
                    "example": 30
                }
            }
        },
...
        "shopping.CartProduct": {
            "type": "object",
            "required": [
                "price_for_quantity",
                "product_id",
                "quantity"
            ],
            "properties": {
                "price_for_quantity": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "shopping.CartProductFullInfo": {
            "type": "object",
            "required": [
                "price_for_quantity",
                "product_id",
                "quantity"
            ],
            "properties": {
                "amount_in_stock": {
                    "type": "number",
                    "example": 120
                },
                "article": {
                    "type": "string",
                    "example": "000123"
                },
                "created_at": {
                    "type": "string"
                },
                "img_url": {
                    "type": "string",
                    "example": "https://google-images.com/some-img123"
                },
                "packages_in_box": {
                    "type": "integer",
                    "example": 50
                },
                "price": {
                    "type": "number",
                    "example": 3.75
                },
                "price_for_quantity": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_title": {
                    "type": "string",
                    "example": "Стакан одноразовий Крафт 110мл"
                },
                "quantity": {
                    "type": "integer"
                },
                "units_in_package": {
                    "type": "integer",
                    "example": 30
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}